+++ src/components/layout/header/About.tsx
@@ -21,14 +21,12 @@ import { DrawerSection, DrawerSectionEntry } from '@/components/DrawerSection';
 const About = () => {
     const { isOpen, onOpen, onClose } = useDisclosure();
 
-    const backendVersion = useAppSelector(
-        (state) => state.organizations.info.version
-    );
+    const backendVersion = useAppSelector((state) => state.me.info.version);
     const orchestratorVersion = useAppSelector(
-        (state) => state.organizations.info.orchestrator_version
+        (state) => state.me.info.orchestrator_version
     );
     const chaincodeVersion = useAppSelector(
-        (state) => state.organizations.info.chaincode_version
+        (state) => state.me.info.chaincode_version
     );
 
     return (