+++ CHANGELOG.md
@@ -10,6 +10,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Changed
 
 -   Upgrade caniuse-lite (#639)
+-   Put current user info into /me module (#641)
 
 ## [0.27.0] - 2022-07-05